requirements.txt
uvicorn
fastapi
pyjwt
sqlalchemy
python-jose[cryptography]
python-multipart
passlib[bcrypt]
mysqlclient
pymysql
